Refactor and re-modularize the whole darn thing
Look, I'm not going to pretend that you're actually going to look through every file in this commit. Just trust that the app is now broken up into common, data, feature, and app modules and properly uses Hilt for dependency injection. I've made the architecture for this app way more complex than probably warrants for a project of this size, but this was a good learning experience, and being honest, I've probably created a good template for other more complex apps.

TODO: Finish polishing app and implement tests